+++ src/ui/main/download_diff.rs
@@ -21,7 +21,7 @@ pub fn download_diff(sender: ComponentSender<App>, progress_bar_input: Sender<Pr
             diff = diff.with_temp_folder(temp);
         }
 
-        let result = diff.install_to(game_path, clone!(
+        let result = diff.install_to(game_path, 1, clone!(
             #[strong]
             sender,